requirements.txt
python-dotenv==1.0.0
random2==1.0.1
openai==0.27.8
gtts==2.2.3
moviepy==1.0.3
curl_cffi
g4f
random_word